Fix Travis CI builds (#2098)
* Temporarily fix Travis CI builds

`nvm install node` will install the latest node version, which is 12 as
of writing. Because of a problem in the resolved node-sass version this
makes the builds fail.

To get green builds until a proper fix for node 12 is introduced we'll
install node 10, which is a LTS release, while 11 will be EOL in June
2019.

@rokumatsumoto is maybe working on a fix, but I think green builds,
especially for third party PRs are an important intermediate step.

https://github.com/rails/webpacker/issues/2077
https://github.com/sass/node-sass/pull/2633
https://github.com/nodejs/nan/issues/849
https://github.com/nodejs/Release

* Lock RuboCop to a version that supports Ruby 2.2

RuboCop 0.69.0 dropped support for Ruby 2.2, making RuboCop fail when it
was installed since it pointed to the git repo instead of a specific
version.

https://github.com/rubocop-hq/rubocop/blob/v0.69.0/CHANGELOG.md#changes

* Allow failures with ruby-head

These might – currently – not be related to Ruby itself, but to the
fact that bundler 2.1.0.pre.1 was installed, which is as of writing
not supported by webpacker (`~> 1.12` is used).

But this is to get the CI builds green again.