When I attempt to run the check_yarn take task when yarn is NOT installed I run into the following issue:
   $ rake webpacker:check_yarn
   bin/rails: No such file or directory - yarn
   rails aborted!
   NoMethodError: undefined method `strip' for nil:NilClass
   Tasks: TOP => webpacker:check_yarn
   (See full trace by running task with --trace)

It looks like lib/tasks/webpacker/check_yarn.rake behaves as expected
when line 6 calls yarn_version = `yarn --version`.strip.  When yarn is
not installed the shell raises an exception and is rescued on line 25.

HOWEVER! When running this in the application, Rails will rescue this exception
and `yarn --version` will return nil.  The tests do not test this adequately
in a Rails environment.

This fix will more elegantly call `which yarn` first and follow the
expected error handling.