Allow liveReload to be configurable in dev server
By default, liveReload for the `webpack_dev_server` config is the inverse of the hmr setting. There may be a usecase where one may want to control both values manually.

This allows liveReload to be configurable through `webpacker.yml` via the `live_reload` option. If not set, then the default of the inverse of hmr is used.

--

As an aside, this also fixes a deprication warning regarding the use of `[hash]` for the output file from the dev server.

```
[DEP_WEBPACK_TEMPLATE_PATH_PLUGIN_REPLACE_PATH_VARIABLES_HASH] DeprecationWarning: [hash] is now [fullhash] (also consider using [chunkhash] or [contenthash], see documentation for details)
```